+++ src/MimesisPlayerEnhancement/Features/WebDashboard/WebDashboardGameAvatarSource.cs
@@ -1,5 +1,4 @@
 using System.Reflection;
-using Steamworks;
 using UnityEngine;
 
 namespace MimesisPlayerEnhancement.Features.WebDashboard